Revert "Merge pull request #8 from PhantomInTheWire/dev"
This reverts commit 4762b9db2f00dcc5ac93ad306276f4a8b92a789d.
