+++ src/__init__.py
@@ -2,9 +2,8 @@ from fastapi import FastAPI
 from .books.routes import book_router
 from fastapi import FastAPI
 from .books.routes import book_router
+from contextlib import asynccontextmanager
 from .auth.routes import auth_router
-from .reviews.routes import review_router
-
 version = 'v1'
 app = FastAPI(
     title='Bookly',
@@ -14,4 +13,3 @@ app = FastAPI(
 
 app.include_router(book_router,prefix=f"/api/{version}/books", tags=['books'])
 app.include_router(auth_router, prefix=f"/api/{version}/auth", tags=['auth'])
-app.include_router(review_router, prefix=f"/api/{version}/reviews", tags=["reviews"])
