+++ src/auth/models.py
@@ -1,0 +1,32 @@
+from sqlmodel import SQLModel, Field, Column
+import sqlalchemy.dialects.postgresql as pg
+import uuid
+from datetime import datetime
+
+class User(SQLModel, table=True):
+    __tablename__ = "user_accounts"
+
+    uid: uuid.UUID = Field(
+        sa_column=Column(
+            pg.UUID,
+            primary_key=True,
+            unique=True,
+            nullable=False,
+            default=uuid.uuid4,
+            info={"description": "Unique identifier for the user account"},
+        )
+    )
+
+    role: str = Field(
+        sa_column=Column(pg.VARCHAR, nullable=False, server_default="user"))
+    username: str
+    first_name: str = Field(nullable=True)
+    last_name: str = Field(nullable=True)
+    is_verified: bool = False
+    email: str = Field(index=True)
+    password_hash: str = Field(exclude=True)
+    created_at: datetime = Field(sa_column=Column(pg.TIMESTAMP, default=datetime.now))
+    updated_at: datetime = Field(sa_column=Column(pg.TIMESTAMP, default=datetime.now))
+
+    def __repr__(self) -> str:
+        return f"<User {self.username}>"
