+++ src/auth/routes.py
@@ -1,6 +1,6 @@
 from fastapi import APIRouter, Depends, status
 from fastapi.responses import JSONResponse
-from .schemas import UserCreateModel, UserModel, UserLoginModel, UserBooksModel
+from .schemas import UserCreateModel, UserModel, UserLoginModel
 from .dependencies import RefreshTokenBearer, AccessTokenBearer, get_current_user
 from .service import UserService
 from src.db.main import get_session
@@ -95,6 +95,6 @@ async def revoke_token(token_details:dict=Depends(AccessTokenBearer())):
         status_code=status.HTTP_200_OK
     )
 
-@auth_router.get("/me", response_model=UserBooksModel)
+@auth_router.get("/me", response_model=UserModel)
 async def get_current_user(user=Depends(get_current_user)):
     return user
