+++ src/auth/service.py
@@ -1,4 +1,4 @@
-from src.db.models import User
+from .models import User
 from .schemas import UserCreateModel
 from .utils import generate_password_hash
 from sqlmodel.ext.asyncio.session import AsyncSession
