+++ src/books/service.py
@@ -1,8 +1,10 @@
 from sqlmodel.ext.asyncio.session import AsyncSession
 from .schemas import BookCreateModel, BookUpdateModel
 from sqlmodel import select, desc
-from src.db.models import Book, User
+from .models import Book
+from datetime import datetime
 import uuid
+from ..auth.models import User
 
 
 class BookService:
