--- src/db/models.py
@@ -1,82 +1,0 @@
-from typing import List
-import sqlalchemy.dialects.postgresql as pg
-from typing import Optional
-from sqlmodel import SQLModel, Field, Relationship, Column
-import uuid
-from datetime import datetime
-
-class User(SQLModel, table=True):
-    __tablename__ = "user_accounts"
-
-    uid: uuid.UUID = Field(
-        sa_column=Column(
-            pg.UUID,
-            primary_key=True,
-            unique=True,
-            nullable=False,
-            default=uuid.uuid4,
-            info={"description": "Unique identifier for the user account"},
-        )
-    )
-
-    role: str = Field(
-        sa_column=Column(pg.VARCHAR, nullable=False, server_default="user"))
-    username: str
-    first_name: str = Field(nullable=True)
-    last_name: str = Field(nullable=True)
-    is_verified: bool = False
-    email: str = Field(index=True)
-    password_hash: str = Field(exclude=True)
-    created_at: datetime = Field(sa_column=Column(pg.TIMESTAMP, default=datetime.now))
-    updated_at: datetime = Field(sa_column=Column(pg.TIMESTAMP, default=datetime.now))
-
-    books: List["Book"] = Relationship(
-        back_populates="user", sa_relationship_kwargs={"lazy": "selectin"}
-    )
-
-    reviews: List["Review"] = Relationship(
-        back_populates="user", sa_relationship_kwargs={"lazy": "selectin"}
-    )
-
-    def __repr__(self) -> str:
-        return f"<User {self.username}>"
-
-class Book(SQLModel, table=True):
-    __tablename__ = "books"
-
-    uid: uuid.UUID = Field(default=None, primary_key=True, nullable=False)
-    title: str
-    author: str
-    publisher: str
-    published_date: str
-    page_count: int
-    language: str
-    user_uid: Optional[uuid.UUID] = Field(default=None, foreign_key="user_accounts.uid")
-    created_at: datetime = Field(default_factory=datetime.now)
-    updated_at: datetime = Field(default_factory=datetime.now)
-
-    user: Optional[User] = Relationship(back_populates="books")
-    reviews: List["Review"] = Relationship(
-        back_populates="book", sa_relationship_kwargs={"lazy": "selectin"}
-    )
-
-    def __repr__(self) -> str:
-        return f"<Book {self.title}>"
-
-
-class Review(SQLModel, table=True):
-    __tablename__ = "reviews"
-    uid: uuid.UUID = Field(
-        sa_column=Column(pg.UUID, nullable=False, primary_key=True, default=uuid.uuid4)
-    )
-    rating: int = Field(lt=6)
-    review_text: str = Field(sa_column=Column(pg.VARCHAR, nullable=False))
-    user_uid: Optional[uuid.UUID] = Field(default=None, foreign_key="user_accounts.uid")
-    book_uid: Optional[uuid.UUID] = Field(default=None, foreign_key="books.uid")
-    created_at: datetime = Field(sa_column=Column(pg.TIMESTAMP, default=datetime.now))
-    updated_at: datetime = Field(sa_column=Column(pg.TIMESTAMP, default=datetime.now))
-    user: Optional[User] = Relationship(back_populates="reviews")
-    book: Optional[Book] = Relationship(back_populates="reviews")
-
-    def __repr__(self):
-        return f"<Review for book {self.book_uid} by user {self.user_uid}>"
